BE_220928_challenge : star challege 등록(완료)

+++ backend/src/main/java/com/ssafy/kirin/entity/CelebChallengeInfo.java
@@ -25,7 +25,7 @@ public class CelebChallengeInfo {
 
     String music;
     String musicTitle;
-    Integer length;
+    Double length;
 
     @Column(name = "stamp_img")
     String stampImg;
